MOVE #102 getMockXY() to logic/mocks/
